TASK: Extract the generation of public CacheControl Headers into separate middleware
This allows the FullPageCache Package to be used in cases where cacheability and lifetime cannot be
extracted from the fusionCache. A possible useCase for that are custom controllers may or may not use
fusion for rendering.

The public lifetime is now calculated from the  header `X-FullPageCache-Lifetime` limited by the Setting `maxPublicCacheTime`.
If the response already contains a `CacheControl` header this is skipped.